requirements.txt
# Legal Document Q&A Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Core LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# PDF Processing
pypdf>=4.0.0

# Vector Store
chromadb>=0.5.0

# Embeddings (optional local alternative)
sentence-transformers>=3.0.0

# UI Framework
streamlit>=1.38.0

# Environment Management
python-dotenv>=1.0.0

# Web Search Fallback
tavily-python>=0.5.0

# Reranking Options
cohere>=5.0.0
flashrank>=0.2.0

# LLM Providers
openai>=1.50.0
groq>=0.11.0

# Utilities
tiktoken>=0.7.0
